feat: add policy details tabs and counters
