+++ web/src/helpers/utils.tsx
@@ -35,6 +35,7 @@ import {
   CHECK_IF_HASH_REGEX,
   SOURCE_LABEL_REGEX,
 } from 'Source/constants';
+import mapValues from 'lodash/mapValues';
 import sum from 'lodash/sum';
 import { ErrorResponse } from 'apollo-link-error';
 import { ApolloError } from '@apollo/client';
@@ -145,6 +146,10 @@ export const getElapsedTime = (unixTimestamp: number) => {
   dayjs.extend(relativeTime);
   return dayjs.unix(unixTimestamp).fromNow();
 };
+
+/** Converts any value of the object that is an array to a comma-separated string */
+export const convertObjArrayValuesToCsv = (obj: { [key: string]: any }) =>
+  mapValues(obj, v => (Array.isArray(v) ? v.join(',') : v));
 
 /**
  * makes sure that it properly formats a JSON struct in order to be properly displayed within the
